requirements.txt
torch
torchvision
torchbearer
numpy
pandas
colour-science
scikit-image